tidied up some of the color manipulation code
