+++ MaterialDesignColors.Wpf/ColorManipulation/Lab.cs
@@ -1,4 +1,6 @@
-﻿namespace MaterialDesignColors.ColorManipulation
+﻿using System;
+
+namespace MaterialDesignColors.ColorManipulation
 {
     public class Lab
     {
@@ -22,7 +24,7 @@
         public const double WhitePointY = 1;
         public const double WhitePointZ = 1.08883;
 
-        public const double t1 = 0.206896552;
+        public static double eCubedRoot = Math.Pow(e, 1.0 / 3);
         public const double k = 24389/27.0;
         public const double e = 216/24389.0;
     }
